+++ lib/Unicode/Char.hs
@@ -27,31 +27,22 @@
 
 module Unicode.Char
     ( module Unicode.Char.General
+    , module Unicode.Char.General.Compat
     , module Unicode.Char.Case
+    , module Unicode.Char.Numeric
     , module Unicode.Char.Normalization
     , module Unicode.Char.Identifiers
 
     -- * Re-export
-    -- ** Numbers
-    , isDigit
-    , isOctDigit
-    , isHexDigit
-    -- ** Subranges
-    , isAscii
-    , isLatin1
-    , isAsciiUpper
-    , isAsciiLower
-    -- ** Numeric representations
     , ord
     , chr
     )
 where
 
-import Data.Char 
-    ( isDigit, isOctDigit, isHexDigit
-    , isAscii, isLatin1, isAsciiUpper, isAsciiLower
-    , chr, ord )
+import Data.Char (chr, ord)
 import Unicode.Char.Case
 import Unicode.Char.General
-import Unicode.Char.Normalization
+import Unicode.Char.General.Compat
 import Unicode.Char.Identifiers
+import Unicode.Char.Numeric
+import Unicode.Char.Normalization