+++ test/Unicode/CharSpec.hs
@@ -10,7 +10,7 @@ import qualified Unicode.Char as UChar
 import Data.Foldable (traverse_)
 import Test.Hspec
 
-{- [NOTE] 
+{- [NOTE]
 These tests may fail if the compiler’s Unicode version
 does not match the version of this package.
 
@@ -28,62 +28,51 @@ does not match the version of this package.
 spec :: Spec
 spec = do
 #ifdef COMPATIBLE_GHC_UNICODE
-  describe "Unicode general categories" do
+  let describe' = describe
 #else
-  let xdescribe' t = before_ (pendingWith "Incompatible GHC Unicode version") . describe t
-  xdescribe' "Unicode general categories" do
+  let describe' t = before_ (pendingWith "Incompatible GHC Unicode version")
+                  . describe t
 #endif
+  describe' "Unicode general categories" do
     it "generalCategory" do
       -- Note: we cannot compare the categories directly, so use 'show'.
-      let sameGC c = show (UChar.generalCategory c) == show (Char.generalCategory c)
-      traverse_ (`shouldSatisfy` sameGC) [minBound..maxBound]
-#ifdef COMPATIBLE_GHC_UNICODE
-  describe "Character classification" do
-#else
-  xdescribe' "Character classification" do
-#endif
+      (show . UChar.generalCategory) `shouldBeEqualTo` (show . Char.generalCategory)
+  describe' "Character classification" do
     it "isAlpha" do
-      let same c = UChar.isAlpha c == Char.isAlpha c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isAlpha `shouldBeEqualTo` Char.isAlpha
     it "isAlphaNum" do
-      let same c = UChar.isAlphaNum c == Char.isAlphaNum c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isAlphaNum `shouldBeEqualTo` Char.isAlphaNum
     it "isControl" do
-      let same c = UChar.isControl c == Char.isControl c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isControl `shouldBeEqualTo` Char.isControl
     it "isLetter" do
-      let same c = UChar.isLetter c == Char.isLetter c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isLetter `shouldBeEqualTo` Char.isLetter
     it "isMark" do
-      let same c = UChar.isMark c == Char.isMark c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isMark `shouldBeEqualTo` Char.isMark
     it "isNumber" do
-      let same c = UChar.isNumber c == Char.isNumber c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isNumber `shouldBeEqualTo` Char.isNumber
     it "isPrint" do
-      let same c = UChar.isPrint c == Char.isPrint c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isPrint `shouldBeEqualTo` Char.isPrint
     it "isPunctuation" do
-      let same c = UChar.isPunctuation c == Char.isPunctuation c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isPunctuation `shouldBeEqualTo` Char.isPunctuation
     it "isSeparator" do
-      let same c = UChar.isSeparator c == Char.isSeparator c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isSeparator `shouldBeEqualTo` Char.isSeparator
     it "isSpace" do
-      let same c = UChar.isSpace c == Char.isSpace c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isSpace `shouldBeEqualTo` Char.isSpace
     it "isSymbol" do
-      let same c = UChar.isSymbol c == Char.isSymbol c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
-#ifdef COMPATIBLE_GHC_UNICODE
-  describe "Case" do
-#else
-  xdescribe' "Case" do
-#endif
-    let xit' t = before_ (pendingWith "Incompatible implementation with Data.Char") . it t
-    xit' "isLower" do
-      let same c = UChar.isLower c == Char.isLower c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
-    xit' "isUpper" do
-      let same c = UChar.isUpper c == Char.isUpper c
-      traverse_ (`shouldSatisfy` same) [minBound..maxBound]
+      UChar.isSymbol `shouldBeEqualTo` Char.isSymbol
+  describe' "Case" do
+    let it' t = before_ (pendingWith "Incompatible implementation with Data.Char")
+              . it t
+    it' "isLower" do
+      UChar.isLower `shouldBeEqualTo` Char.isLower
+    it' "isUpper" do
+      UChar.isUpper `shouldBeEqualTo` Char.isUpper
+  where
+    shouldBeEqualTo
+        :: forall a b. (Bounded a, Enum a, Show a, Eq b, Show b)
+        => (a -> b)
+        -> (a -> b)
+        -> IO ()
+    shouldBeEqualTo f g =
+        let same x = f x == g x
+        in traverse_ (`shouldSatisfy` same) [minBound..maxBound]